add fonts and continue decompose widget
