+++ lib/features/sber_settings/view/view.dart
@@ -1,1 +1,1 @@
-// TODO Implement this library.
+export '../view/sber_settings_screen.dart';
