+++ pubspec.yaml
@@ -74,17 +74,15 @@ flutter:
   # "family" key with the font family name, and a "fonts" key with a
   # list giving the asset and other descriptors for the font. For
   # example:
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
-  #
-  # For details regarding fonts from package dependencies,
-  # see https://flutter.dev/custom-fonts/#from-packages
+  fonts:
+    - family: SF Pro Text
+      fonts:
+        - asset: lib/ui/fonts/SFProText/SFProText_Regular.ttf
+        - asset: lib/ui/fonts/SFProDisplay/SFProText_Medium.ttf
+        - asset: lib/ui/fonts/SFProText/SFProText_Bold.ttf
+          weight: 800
+    - family: SF Pro Display
+      fonts:
+        - asset: lib/ui/fonts/SFProDisplay/SFProDisplay_Regular.otf
+        - asset: lib/ui/fonts/SFProDisplay/SFProDisplay_Bold.otf
+          weight: 700
